refactor: extract StructureBuilder class (#25)
Move all Pass 1 (structure building) logic from resolver.ts into a
dedicated StructureBuilder class. This encapsulates buildResObj,
applyField, astToResolverValue, and relativize functions as class
methods, with opts and loader as instance fields.
