+++ src/internal/resolver/resolver.ts
@@ -1,6 +1,6 @@
 import { ResolveError } from '../../errors.js'
 import type { HoconValue } from '../../value.js'
-import type { AstNode, AstField } from '../parser/ast.js'
+import type { AstNode } from '../parser/ast.js'
 import {
   type SubstPlaceholder,
   type AppendPlaceholder,
@@ -12,16 +12,14 @@ import {
   isConcat,
   isAppend,
   isResObj,
-  makeResObj,
 } from './types.js'
 import {
   deepMergeHoconValues,
-  deepMergeResObjInto,
   lookupPath,
   lookupResObj,
   parseSubstPath,
 } from './utils.js'
-import { IncludeLoader } from './include-loader.js'
+import { StructureBuilder } from './structure-builder.js'
 
 export type { ResolveOptions } from './types.js'
 
@@ -29,7 +27,7 @@ export type { ResolveOptions } from './types.js'
 
 export function resolve(ast: AstNode, opts: ResolveOptions): HoconValue {
   // Pass 1
-  const root = buildResObj(ast, opts, [])
+  const root = new StructureBuilder(opts).build(ast)
   // Pass 2
   const resolving = new Set<string>()
   const resolvedCache = new Map<string, HoconValue>()
@@ -38,152 +36,11 @@ export function resolve(ast: AstNode, opts: ResolveOptions): HoconValue {
 
 export async function resolveAsync(ast: AstNode, opts: ResolveOptions): Promise<HoconValue> {
   // Pass 1 (async — awaits file reads for includes)
-  const root = await buildResObjAsync(ast, opts, [])
+  const root = await new StructureBuilder(opts).buildAsync(ast)
   // Pass 2 (sync — no I/O needed for substitution resolution)
   const resolving = new Set<string>()
   const resolvedCache = new Map<string, HoconValue>()
   return resolveResObj(root, root, resolving, resolvedCache, opts)
-}
-
-// ---- Pass 1: structure building ----
-
-function buildResObj(ast: AstNode, opts: ResolveOptions, pathPrefix: string[]): ResObj {
-  if (ast.kind !== 'object') {
-    throw new ResolveError('root AST must be an object', '', ast.pos.line, ast.pos.col)
-  }
-  const loader = new IncludeLoader(opts)
-  loader.onBuildResObj = (a, o) => buildResObj(a, o, [])
-  // onBuildResObjAsync not needed for sync path but set a stub to satisfy the type
-  loader.onBuildResObjAsync = async (a, o) => buildResObjAsync(a, o, [])
-  const obj = makeResObj()
-  for (const field of ast.fields) {
-    applyField(obj, field, opts, pathPrefix, loader)
-  }
-  return obj
-}
-
-function applyField(obj: ResObj, field: AstField, opts: ResolveOptions, pathPrefix: string[], loader: IncludeLoader): void {
-  // include directive: key is empty, value is include node
-  if (field.key.length === 0 && field.value.kind === 'include') {
-    // Included files are parsed at their own root (pathPrefix=[]),
-    // then relativized to the current scope's prefix.
-    const included = loader.load(field.value.path, field.value.required)
-    if (pathPrefix.length > 0) {
-      relativizeResObj(included, pathPrefix.join('.'), pathPrefix.length)
-    }
-    deepMergeResObjInto(obj, included)
-    return
-  }
-
-  const [head, ...tail] = field.key
-  if (!head) return
-
-  if (tail.length > 0) {
-    // Nested key: server.host = "x" → create synthetic object AST
-    const syntheticAst: AstNode = {
-      kind: 'object',
-      fields: [{ key: tail, value: field.value, append: field.append, pos: field.pos }],
-      pos: field.pos,
-    }
-    applyField(obj, { key: [head], value: syntheticAst, append: false, pos: field.pos }, opts, pathPrefix, loader)
-    return
-  }
-
-  const childPrefix = [...pathPrefix, head]
-
-  if (field.append) {
-    // +=: append elem to existing array (or start from [])
-    const existing: ResolverValue = obj.fields.get(head) ?? ({ kind: 'array', items: [] } satisfies HoconValue)
-    obj.priorValues.set(head, existing)
-    obj.fields.set(head, {
-      _kind: 'append-placeholder',
-      existing,
-      elem: astToResolverValue(field.value, opts, childPrefix),
-    })
-    return
-  }
-
-  // Normal assignment
-  const existing = obj.fields.get(head)
-  const newVal = astToResolverValue(field.value, opts, childPrefix)
-
-  // Save prior value for self-referential substitution resolution
-  if (existing !== undefined) {
-    obj.priorValues.set(head, existing)
-  }
-
-  // Deep merge if both are ResObj
-  if (existing !== undefined && isResObj(existing) && isResObj(newVal)) {
-    deepMergeResObjInto(existing, newVal)
-    // existing already in fields — no re-set needed
-    return
-  }
-
-  obj.fields.set(head, newVal)
-}
-
-function astToResolverValue(ast: AstNode, opts: ResolveOptions, pathPrefix: string[]): ResolverValue {
-  switch (ast.kind) {
-    case 'scalar': {
-      const sv: HoconValue = { kind: 'scalar', value: ast.value }
-      if (ast._separator) separatorValues.add(sv)
-      return sv
-    }
-    case 'array':
-      return { kind: 'array', items: ast.items.map(i => astToResolverValue(i, opts, pathPrefix) as HoconValue) }
-    case 'object': {
-      const inner = buildResObj(ast, opts, pathPrefix)
-      return inner
-    }
-    case 'subst':
-      return { _kind: 'subst-placeholder', path: ast.path, optional: ast.optional, line: ast.pos.line, col: ast.pos.col, prefixLen: 0 }
-    case 'concat':
-      return { _kind: 'concat-placeholder', nodes: ast.nodes.map(n => astToResolverValue(n, opts, pathPrefix)) }
-    case 'include':
-      return { kind: 'scalar', value: null } // handled by applyField; should not reach here
-  }
-}
-
-// ---- Relativize substitution paths for nested includes ----
-
-function relativizeSubstPaths(val: ResolverValue, prefix: string, prefixSegmentCount: number): void {
-  if (isSubst(val)) {
-    val.path = `${prefix}.${val.path}`
-    val.prefixLen += prefixSegmentCount
-    return
-  }
-  if (isConcat(val)) {
-    for (const node of val.nodes) {
-      relativizeSubstPaths(node, prefix, prefixSegmentCount)
-    }
-    return
-  }
-  if (isAppend(val)) {
-    relativizeSubstPaths(val.existing, prefix, prefixSegmentCount)
-    relativizeSubstPaths(val.elem, prefix, prefixSegmentCount)
-    return
-  }
-  if (isResObj(val)) {
-    relativizeResObj(val, prefix, prefixSegmentCount)
-    return
-  }
-  // HoconValue arrays may contain substitutions inside items (shouldn't happen
-  // in practice since arrays are built from astToResolverValue, but be safe)
-  const hv = val as HoconValue
-  if (hv.kind === 'array') {
-    for (const item of hv.items) {
-      relativizeSubstPaths(item as ResolverValue, prefix, prefixSegmentCount)
-    }
-  }
-}
-
-function relativizeResObj(obj: ResObj, prefix: string, prefixSegmentCount: number): void {
-  for (const val of obj.fields.values()) {
-    relativizeSubstPaths(val, prefix, prefixSegmentCount)
-  }
-  for (const val of obj.priorValues.values()) {
-    relativizeSubstPaths(val, prefix, prefixSegmentCount)
-  }
 }
 
 // ---- Pass 2: substitution resolution ----
@@ -428,101 +285,4 @@ function resolveAppend(
   const items: HoconValue[] = existing.kind === 'array' ? [...existing.items] : [existing]
   if (elem !== undefined) items.push(elem)
   return { kind: 'array', items }
-}
-
-// ---- Async Pass 1 helpers (mirror sync versions but await file reads) ----
-
-async function buildResObjAsync(ast: AstNode, opts: ResolveOptions, pathPrefix: string[]): Promise<ResObj> {
-  if (ast.kind !== 'object') {
-    throw new ResolveError('root AST must be an object', '', ast.pos.line, ast.pos.col)
-  }
-  const loader = new IncludeLoader(opts)
-  loader.onBuildResObj = (a, o) => buildResObj(a, o, [])
-  loader.onBuildResObjAsync = async (a, o) => buildResObjAsync(a, o, [])
-  const obj = makeResObj()
-  for (const field of ast.fields) {
-    await applyFieldAsync(obj, field, opts, pathPrefix, loader)
-  }
-  return obj
-}
-
-async function applyFieldAsync(obj: ResObj, field: AstField, opts: ResolveOptions, pathPrefix: string[], loader: IncludeLoader): Promise<void> {
-  if (field.key.length === 0 && field.value.kind === 'include') {
-    const included = await loader.loadAsync(field.value.path, field.value.required)
-    if (pathPrefix.length > 0) {
-      relativizeResObj(included, pathPrefix.join('.'), pathPrefix.length)
-    }
-    deepMergeResObjInto(obj, included)
-    return
-  }
-
-  const [head, ...tail] = field.key
-  if (!head) return
-
-  if (tail.length > 0) {
-    const syntheticAst: AstNode = {
-      kind: 'object',
-      fields: [{ key: tail, value: field.value, append: field.append, pos: field.pos }],
-      pos: field.pos,
-    }
-    await applyFieldAsync(obj, { key: [head], value: syntheticAst, append: false, pos: field.pos }, opts, pathPrefix, loader)
-    return
-  }
-
-  const childPrefix = [...pathPrefix, head]
-
-  if (field.append) {
-    const existing: ResolverValue = obj.fields.get(head) ?? ({ kind: 'array', items: [] } satisfies HoconValue)
-    obj.priorValues.set(head, existing)
-    obj.fields.set(head, {
-      _kind: 'append-placeholder',
-      existing,
-      elem: await astToResolverValueAsync(field.value, opts, childPrefix),
-    })
-    return
-  }
-
-  const existing = obj.fields.get(head)
-  const newVal = await astToResolverValueAsync(field.value, opts, childPrefix)
-
-  if (existing !== undefined) {
-    obj.priorValues.set(head, existing)
-  }
-
-  if (existing !== undefined && isResObj(existing) && isResObj(newVal)) {
-    deepMergeResObjInto(existing, newVal)
-    return
-  }
-
-  obj.fields.set(head, newVal)
-}
-
-async function astToResolverValueAsync(ast: AstNode, opts: ResolveOptions, pathPrefix: string[]): Promise<ResolverValue> {
-  switch (ast.kind) {
-    case 'scalar': {
-      const sv: HoconValue = { kind: 'scalar', value: ast.value }
-      if (ast._separator) separatorValues.add(sv)
-      return sv
-    }
-    case 'array': {
-      const items = []
-      for (const i of ast.items) {
-        items.push(await astToResolverValueAsync(i, opts, pathPrefix) as HoconValue)
-      }
-      return { kind: 'array', items }
-    }
-    case 'object':
-      return await buildResObjAsync(ast, opts, pathPrefix)
-    case 'subst':
-      return { _kind: 'subst-placeholder', path: ast.path, optional: ast.optional, line: ast.pos.line, col: ast.pos.col, prefixLen: 0 }
-    case 'concat': {
-      const nodes = []
-      for (const n of ast.nodes) {
-        nodes.push(await astToResolverValueAsync(n, opts, pathPrefix))
-      }
-      return { _kind: 'concat-placeholder', nodes }
-    }
-    case 'include':
-      return { kind: 'scalar', value: null }
-  }
 }
